Move axes_grid tests to axes_grid1
