+++ lib/mpl_toolkits/tests/baseline_images/test_axes_grid1/imagegrid_cbar_mode.png
